requirements.txt
asgiref==3.7.2
cfgv==3.3.1
distlib==0.3.6
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
exceptiongroup==1.1.2
filelock==3.12.2
identify==2.5.24
inflection==0.5.1
iniconfig==2.0.0
nodeenv==1.8.0
packaging==23.1
platformdirs==3.8.1
pluggy==1.2.0
pre-commit==3.3.3
psycopg2-binary==2.9.6
PyJWT==2.7.0
pytest==7.4.0
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
virtualenv==20.24.0
